Fix build issues under Clang
Most of these changes are trivial to satisfy Clang's warnings. Clang did
find one small bug in the move constructor for
Core::Debug::DebugMessage, where processName and threadName were not
moved over as they should have been.

Adds Clang v3.4-3.7 as supported compilers to the README and adds
instructions to CLANG.md.

Towards #9: port to clang

+++ Examples/ReconfigureTest.cc
@@ -65,7 +65,7 @@ class OptionParser {
                     fatal = true;
                     break;
                 case 's':
-                    srand(atoi(optarg));
+                    srand(static_cast<unsigned int>(atoi(optarg)));
                     break;
                 case '?':
                 default:
@@ -199,10 +199,11 @@ main(int argc, char** argv)
     while (true) {
         Configuration newConfiguration;
         uint64_t desiredServers =
-            (rand() % fullConfiguration.size()) + 1; // NOLINT
+            (uint64_t(rand()) % fullConfiguration.size()) + 1; // NOLINT
         Configuration remainingServers = fullConfiguration;
         for (uint64_t i = 0; i < desiredServers; ++i) {
-            uint64_t j = (rand() % remainingServers.size()); // NOLINT
+            uint64_t j =
+                (uint64_t(rand()) % remainingServers.size()); // NOLINT
             newConfiguration.push_back(remainingServers.at(j));
             std::swap(remainingServers.at(j),
                       remainingServers.back());
